fix: sending big chunk of data
Manually send chunks until the buffer is empty. By default, `asio::write` will fail during the process, not as expected. Retry when sending a chunk failed, up to a limit, by default 10. Execute an optional error callback when the limit is reached. Set the socket in disconnected state when the end-side was closed while trying to send a message.